+++ d60.Cirqus.MongoDb/Views/New/NewMongoDbViewManager.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using d60.Cirqus.Events;
 using d60.Cirqus.Extensions;
 using d60.Cirqus.Logging;
-using d60.Cirqus.Numbers;
 using d60.Cirqus.Views.ViewManagers;
 using d60.Cirqus.Views.ViewManagers.Locators;
 using d60.Cirqus.Views.ViewManagers.New;
@@ -20,7 +20,7 @@ namespace d60.Cirqus.MongoDb.Views.New
     {
         const string LowWatermarkDocId = "__low_watermark__";
         const string LowWatermarkPropertyName = "LastGlobalSequenceNumber";
-        const int DefaultLowWatermark = -1;
+        const long DefaultLowWatermark = -1;
 
         readonly ViewDispatcherHelper<TViewInstance> _dispatcherHelper = new ViewDispatcherHelper<TViewInstance>();
         readonly MongoCollection<TViewInstance> _viewCollection;
@@ -69,7 +69,7 @@ namespace d60.Cirqus.MongoDb.Views.New
 
         public long GetLowWatermark(bool canGetFromCache = true)
         {
-            if (canGetFromCache)
+            if (canGetFromCache && false)
             {
                 return GetLowWatermarkFromMemory()
                        ?? GetLowWatermarkFromPersistentCache()
@@ -82,7 +82,7 @@ namespace d60.Cirqus.MongoDb.Views.New
                        ?? GetDefaultLowWatermark();
         }
 
-        static int GetDefaultLowWatermark()
+        static long GetDefaultLowWatermark()
         {
             return DefaultLowWatermark;
         }
@@ -96,17 +96,17 @@ namespace d60.Cirqus.MongoDb.Views.New
             Interlocked.Exchange(ref _cachedLowWatermark, DefaultLowWatermark);
         }
 
-        public async Task WaitUntilDispatched(CommandProcessingResult result, TimeSpan timeout)
+        public async Task WaitUntilProcessed(CommandProcessingResult result, TimeSpan timeout)
         {
             if (!result.EventsWereEmitted) return;
 
             var mostRecentGlobalSequenceNumber = result.GlobalSequenceNumbersOfEmittedEvents.Max();
 
-            var waitStartTime = DateTime.UtcNow;
+            var stopwatch = Stopwatch.StartNew();
 
             while (GetLowWatermark(canGetFromCache: false) < mostRecentGlobalSequenceNumber)
             {
-                if (DateTime.UtcNow - waitStartTime > timeout)
+                if (stopwatch.Elapsed > timeout)
                 {
                     throw new TimeoutException(string.Format("View for {0} did not catch up to {1} within {2} timeout!",
                         typeof(TViewInstance), mostRecentGlobalSequenceNumber, timeout));
@@ -181,7 +181,10 @@ namespace d60.Cirqus.MongoDb.Views.New
         {
             var value = Interlocked.Read(ref _cachedLowWatermark);
 
-            return value != DefaultLowWatermark ? value : default(long?);
+            if (value != DefaultLowWatermark) 
+                return value;
+            
+            return null;
         }
 
         long? GetLowWatermarkFromPersistentCache()
@@ -216,7 +219,7 @@ namespace d60.Cirqus.MongoDb.Views.New
                 return lowWatermark;
             }
 
-            return default(long?);
+            return null;
         }
     }
 }